+++ sdk/baremetalinfrastructure/azure-mgmt-baremetalinfrastructure/generated_samples/azure_bare_metal_instances_delete.py
@@ -16,7 +16,7 @@ from azure.mgmt.baremetalinfrastructure import BareMetalInfrastructureClient
     pip install azure-identity
     pip install azure-mgmt-baremetalinfrastructure
 # USAGE
-    python azure_bare_metal_operations_list.py
+    python azure_bare_metal_instances_delete.py
 
     Before run the sample, please set the values of the client ID, tenant ID and client secret
     of the AAD application as environment variables: AZURE_CLIENT_ID, AZURE_TENANT_ID,
@@ -28,14 +28,15 @@ from azure.mgmt.baremetalinfrastructure import BareMetalInfrastructureClient
 def main():
     client = BareMetalInfrastructureClient(
         credential=DefaultAzureCredential(),
-        subscription_id="SUBSCRIPTION_ID",
+        subscription_id="f0f4887f-d13c-4943-a8ba-d7da28d2a3fd",
     )
 
-    response = client.operations.list()
-    for item in response:
-        print(item)
+    client.azure_bare_metal_instances.delete(
+        resource_group_name="myResourceGroup",
+        azure_bare_metal_instance_name="myAzureBareMetalInstance",
+    )
 
 
-# x-ms-original-file: specification/baremetalinfrastructure/resource-manager/Microsoft.BareMetalInfrastructure/BareMetalInfrastructure/preview/2024-08-01-preview/examples/AzureBareMetalOperations_List.json
+# x-ms-original-file: specification/baremetalinfrastructure/resource-manager/Microsoft.BareMetalInfrastructure/BareMetalInfrastructure/preview/2024-08-01-preview/examples/AzureBareMetalInstances_Delete.json
 if __name__ == "__main__":
     main()